Fix the bug
tearDown() is called after each testing function, and gave error after test_read_sudoku() as no output file was created.

+++ tests/test_sudoku_io.py
@@ -21,7 +21,8 @@ class TestSudokuIO(unittest.TestCase):
 		self.output_path = os.path.join('tests', 'test_data', 'example_out.suk')
 
 	def tearDown(self):
-		os.remove(self.output_path)
+		if os.path.exists(self.output_path):
+			os.remove(self.output_path)
 
 	def test_read_sudoku(self):
 		read_array = read_sudoku(self.input_path)
